requirements.txt
firebase-admin
kafka-python
python-dotenv